refactor(core): decompose query and registry responsibilities
